Adopt client tests to the dataclasses Remove useless responses validity checks Run only unit tests on CI
