+++ .github/workflows/tests.yml
@@ -34,7 +34,7 @@ jobs:
       continue-on-error: true
 
     - name: Test with pytest
-      run: poetry run pytest tests
+      run: poetry run pytest -m unit tests
       env:
         LIDO_CONTRACT_ADDRESS: "0x1643E812aE58766192Cf7D2Cf9567dF2C37e9B7F"
